move the terminal's platform code into a tty module
Terminal held a raw handle, the termios it has to put back, the wake pipe
and a SIGWINCH slot. All of that now lives behind Tty, and Terminal asks it
to read, wait, wake, watch for resizes and report its size. Restoring the
saved mode moves with it, and still happens after Terminal writes its exit
sequences because the field drops last.

Code motion, no behaviour change.
